Reorganize KafkaAdminAPITest

+++ src/test/java/io/managed/services/test/KafkaAdminAPITest.java
@@ -5,6 +5,7 @@ import io.managed.services.test.client.kafkaadminapi.KafkaAdminAPI;
 import io.managed.services.test.client.kafkaadminapi.KafkaAdminAPIUtils;
 import io.managed.services.test.client.kafkaadminapi.Topic;
 import io.managed.services.test.client.serviceapi.CreateKafkaPayload;
+import io.managed.services.test.client.serviceapi.KafkaResponse;
 import io.managed.services.test.client.serviceapi.ServiceAPI;
 import io.managed.services.test.client.serviceapi.ServiceAPIUtils;
 import io.managed.services.test.framework.TestTag;
@@ -42,17 +43,16 @@ public class KafkaAdminAPITest extends TestBase {
     private static final Logger LOGGER = LogManager.getLogger(ServiceAPITest.class);
 
     KafkaAdminAPI kafkaAdminAPI;
-    ServiceAPI api;
+    ServiceAPI serviceAPI;
+    KafkaResponse kafka;
     Topic topic;
     String group;
-    String bootstrapServerHost;
-
 
     static final String KAFKA_INSTANCE_NAME = "mk-e2e-kaa-" + Environment.KAFKA_POSTFIX_NAME;
     static final String TEST_TOPIC_NAME = "test-api-topic-1";
-    static final String TEST_NOT_EXISTING_TOPIC_NAME = "test-api-topic-notExist";
+    static final String TEST_NOT_EXISTING_TOPIC_NAME = "test-api-topic-not-exist";
 
-    static final String TEST_GROUP_NAME = "strimzi-canary-group";
+    static final String TEST_ACTIVE_GROUP_NAME = "strimzi-canary-group";
     static final String TEST_NOT_EXISTING_GROUP_NAME = "not-existing-group";
 
 
@@ -60,7 +60,7 @@ public class KafkaAdminAPITest extends TestBase {
     @BeforeAll
     void bootstrap(Vertx vertx, VertxTestContext context) {
         ServiceAPIUtils.serviceAPI(vertx)
-                .onSuccess(a -> api = a)
+                .onSuccess(a -> serviceAPI = a)
                 .onComplete(context.succeedingThenComplete());
     }
 
@@ -68,31 +68,32 @@ public class KafkaAdminAPITest extends TestBase {
     @AfterAll
     void deleteKafkaInstance(VertxTestContext context) {
 
-        ServiceAPIUtils.deleteKafkaByNameIfExists(api, KAFKA_INSTANCE_NAME)
+        ServiceAPIUtils.deleteKafkaByNameIfExists(serviceAPI, KAFKA_INSTANCE_NAME)
                 .onComplete(context.succeedingThenComplete());
     }
 
-    void assertRestAPI() {
-        assumeTrue(kafkaAdminAPI != null, "rest API is null because the createKafkaUsingServiceAPI has failed");
+    void assertKafkaAdminAPI() {
+        assumeTrue(kafkaAdminAPI != null, "kafkaAdminAPI is null because the testConnectKafkaAdminAPI has failed");
     }
 
     void assertTopic() {
-        assumeTrue(topic != null, "topic is null because the testCreateTopic has failed to create the topic on the Kafka instance");
+        assumeTrue(topic != null, "topic is null because the testCreateTopic has failed");
     }
 
-    void assertAPI() {
-        assumeTrue(api != null, "api is null because the bootstrap has failed");
+    void assertServiceAPI() {
+        assumeTrue(serviceAPI != null, "serviceAPI is null because the bootstrap has failed");
     }
 
-    void assertConnectedToRunningKafkaInstance() {
-        assumeTrue(bootstrapServerHost != null, "Failed to connect to Kafka within Timeout Period (8minutes 20 seconds)");
+    void assertKafka() {
+        assumeTrue(kafka != null, "kafka is null because the testCreateKafkaInstance has failed");
     }
 
     @Test
     @Order(1)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
     void testCreateKafkaInstance(Vertx vertx, VertxTestContext context) {
-        assertAPI();
+        assertServiceAPI();
+
         // Create Kafka Instance
         CreateKafkaPayload kafkaPayload = new CreateKafkaPayload();
         // add postfix to the name based on owner
@@ -102,18 +103,21 @@ public class KafkaAdminAPITest extends TestBase {
         kafkaPayload.region = "us-east-1";
 
         LOGGER.info("create kafka instance: {}", kafkaPayload.name);
-        api.createKafka(kafkaPayload, true)
-                .compose(k -> waitUntilKafkaIsReady(vertx, api, k.id))
-                .onSuccess(k -> bootstrapServerHost = k.bootstrapServerHost)
+        serviceAPI.createKafka(kafkaPayload, true)
+                .compose(k -> waitUntilKafkaIsReady(vertx, serviceAPI, k.id))
+                .onSuccess(k -> kafka = k)
                 .onComplete(context.succeedingThenComplete());
     }
 
     @Test
     @Order(2)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
-    void connectKafkaAdminAPI(Vertx vertx, VertxTestContext context) {
-        assertAPI();
-        assertConnectedToRunningKafkaInstance();
+    void testConnectKafkaAdminAPI(Vertx vertx, VertxTestContext context) {
+        assertServiceAPI();
+        assertKafka();
+
+        var bootstrapServerHost = kafka.bootstrapServerHost;
+
         KafkaAdminAPIUtils.kafkaAdminAPI(vertx, bootstrapServerHost)
                 .onSuccess(restApiResponse -> kafkaAdminAPI = restApiResponse)
                 .onFailure(msg -> System.out.println(msg.getMessage()))
@@ -125,10 +129,10 @@ public class KafkaAdminAPITest extends TestBase {
     @Order(3)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
     void testCreateTopic(VertxTestContext context) {
-        assertConnectedToRunningKafkaInstance();
-        assertRestAPI();
+        assertKafkaAdminAPI();
+
         kafkaAdminAPI.getTopicByName(TEST_TOPIC_NAME)
-                .compose(r -> Future.failedFuture("Getting test-topic should fail in 1st test"))
+                .compose(r -> Future.failedFuture("Getting test-topic should fail because the topic shouldn't exists"))
                 .recover(throwable -> {
                     if ((throwable instanceof ResponseException) && (((ResponseException) throwable).response.statusCode() == HttpURLConnection.HTTP_NOT_FOUND)) {
                         LOGGER.info("Topic not found : {}", TEST_TOPIC_NAME);
@@ -147,7 +151,7 @@ public class KafkaAdminAPITest extends TestBase {
     @Order(4)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
     void testCreateExistingTopic(VertxTestContext context) {
-        assertRestAPI();
+        assertKafkaAdminAPI();
         assertTopic();
 
         KafkaAdminAPIUtils.createDefaultTopic(kafkaAdminAPI, TEST_TOPIC_NAME)
@@ -168,11 +172,11 @@ public class KafkaAdminAPITest extends TestBase {
     @Order(4)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
     void testGetTopicByName(VertxTestContext context) {
-        assertRestAPI();
+        assertKafkaAdminAPI();
         assertTopic();
 
         kafkaAdminAPI.getTopicByName(TEST_TOPIC_NAME)
-                .onSuccess(topic -> context.verify(() -> assertEquals(TEST_TOPIC_NAME, topic.name)))
+                .onSuccess(t -> context.verify(() -> assertEquals(TEST_TOPIC_NAME, t.name)))
                 .onComplete(context.succeedingThenComplete());
     }
 
@@ -180,8 +184,7 @@ public class KafkaAdminAPITest extends TestBase {
     @Order(4)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
     void testGetNotExistingTopicByName(VertxTestContext context) {
-        assertConnectedToRunningKafkaInstance();
-        assertRestAPI();
+        assertKafkaAdminAPI();
 
         kafkaAdminAPI.getTopicByName(TEST_NOT_EXISTING_TOPIC_NAME)
                 .compose(r -> Future.failedFuture("Get none existing topic should fail"))
@@ -195,19 +198,22 @@ public class KafkaAdminAPITest extends TestBase {
                     return Future.failedFuture(throwable);
                 })
                 .onComplete(context.succeedingThenComplete());
-
     }
 
 
     @Test
     @Order(4)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
-    void tetGetTopics(VertxTestContext context) {
-        assertRestAPI();
+    void tetGetAllTopics(VertxTestContext context) {
+        assertKafkaAdminAPI();
         assertTopic();
+
         kafkaAdminAPI.getAllTopics()
                 .onSuccess(topics -> context.verify(() -> {
-                    List<Topic> filteredTopics = topics.items.stream().filter(k -> k.name.equals(TEST_TOPIC_NAME) || k.name.equals("strimzi-canary")).collect(Collectors.toList());
+                    List<Topic> filteredTopics = topics.items.stream()
+                            .filter(k -> k.name.equals(TEST_TOPIC_NAME) || k.name.equals("strimzi-canary"))
+                            .collect(Collectors.toList());
+
                     assertEquals(2, filteredTopics.size());
                 }))
                 .onComplete(context.succeedingThenComplete());
@@ -218,8 +224,9 @@ public class KafkaAdminAPITest extends TestBase {
     @Order(5)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
     void testDeleteTopicByName(VertxTestContext context) {
-        assertConnectedToRunningKafkaInstance();
-        assertRestAPI();
+        assertKafkaAdminAPI();
+        assertTopic();
+
         kafkaAdminAPI.deleteTopicByName(TEST_TOPIC_NAME)
                 .compose(r -> kafkaAdminAPI.getTopicByName(TEST_TOPIC_NAME))
                 .compose(r -> Future.failedFuture("Getting test-topic should fail due to topic being deleted in current test"))
@@ -239,8 +246,8 @@ public class KafkaAdminAPITest extends TestBase {
     @Order(5)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
     void testDeleteNotExistingTopicByName(VertxTestContext context) {
-        assertConnectedToRunningKafkaInstance();
-        assertRestAPI();
+        assertKafkaAdminAPI();
+
         kafkaAdminAPI.deleteTopicByName(TEST_NOT_EXISTING_TOPIC_NAME)
                 .compose(__ -> Future.failedFuture("Deleting not existing topic should result in HTTP_NOT_FOUND response"))
                 .recover(throwable -> {
@@ -257,11 +264,12 @@ public class KafkaAdminAPITest extends TestBase {
     @Test
     @Order(3)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
-    void testGetGroups(VertxTestContext context) {
-        assertConnectedToRunningKafkaInstance();
-        assertRestAPI();
+    void testGetAllGroups(VertxTestContext context) {
+        assertKafkaAdminAPI();
+
         kafkaAdminAPI.getAllGroups()
                 .onSuccess(groupResponse -> context.verify(() -> {
+                    // TODO: create a consumer and assert that the created consumer group exists
                     int groupsCount = groupResponse.length;
                     assertTrue(groupsCount >= 1);
                 }))
@@ -272,11 +280,11 @@ public class KafkaAdminAPITest extends TestBase {
     @Order(3)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
     void testGetGroupByName(VertxTestContext context) {
-        assertConnectedToRunningKafkaInstance();
-        assertRestAPI();
-        kafkaAdminAPI.getGroupByName(TEST_GROUP_NAME)
+        assertKafkaAdminAPI();
+
+        kafkaAdminAPI.getGroupByName(TEST_ACTIVE_GROUP_NAME)
                 .onSuccess(groupResponse -> context.verify(() -> {
-                    assertEquals(groupResponse.id, TEST_GROUP_NAME);
+                    assertEquals(groupResponse.id, TEST_ACTIVE_GROUP_NAME);
                     assertEquals("STABLE", groupResponse.state);
                     group = groupResponse.id;
                 }))
@@ -287,8 +295,8 @@ public class KafkaAdminAPITest extends TestBase {
     @Order(3)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
     void testGetNotExistingGroupByName(VertxTestContext context) {
-        assertConnectedToRunningKafkaInstance();
-        assertRestAPI();
+        assertKafkaAdminAPI();
+
         kafkaAdminAPI.getGroupByName(TEST_NOT_EXISTING_GROUP_NAME)
                 .onSuccess(groupResponse -> context.verify(() -> assertEquals("DEAD", groupResponse.state)))
                 .onComplete(context.succeedingThenComplete());
@@ -297,16 +305,15 @@ public class KafkaAdminAPITest extends TestBase {
     @Test
     @Order(4)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
-    void testDeleteGroupByName(VertxTestContext context) {
-        assertConnectedToRunningKafkaInstance();
-        assertRestAPI();
+    void testDeleteActiveGroupByName(VertxTestContext context) {
+        assertKafkaAdminAPI();
 
-        kafkaAdminAPI.deleteGroupByName(TEST_GROUP_NAME)
-                .compose(r -> Future.failedFuture("Deleting existing not empty grop by name"))
+        kafkaAdminAPI.deleteGroupByName(TEST_ACTIVE_GROUP_NAME)
+                .compose(r -> Future.failedFuture("Deleting existing not empty group by name"))
                 .recover(throwable -> {
                     if (throwable instanceof ResponseException) {
                         if (((ResponseException) throwable).response.statusCode() == HttpURLConnection.HTTP_UNAUTHORIZED) {
-                            LOGGER.info("Group isn't empty, thus cannot be deleted : {}", TEST_GROUP_NAME);
+                            LOGGER.info("Group isn't empty, thus cannot be deleted : {}", TEST_ACTIVE_GROUP_NAME);
                             return Future.succeededFuture();
                         }
                     }
@@ -319,8 +326,8 @@ public class KafkaAdminAPITest extends TestBase {
     @Order(3)
     @Timeout(value = 15, timeUnit = TimeUnit.MINUTES)
     void testDeleteNotExistingGroupByName(VertxTestContext context) {
-        assertConnectedToRunningKafkaInstance();
-        assertRestAPI();
+        assertKafkaAdminAPI();
+
         kafkaAdminAPI.deleteGroupByName(TEST_NOT_EXISTING_GROUP_NAME)
                 .compose(r -> Future.failedFuture("Deleting group by not existing name"))
                 .recover(throwable -> {
@@ -333,6 +340,7 @@ public class KafkaAdminAPITest extends TestBase {
                     return Future.failedFuture(throwable);
                 })
                 .onComplete(context.succeedingThenComplete());
-
     }
+
+    // TODO: Tests delete group for real
 }
